Update Automate-Module.psm1
Improve Uninstall-Automate
Query for multiple existing uninstall packages to be removed during manual gutting. 
Was getting Exit Code 1638 - "Another version of this product is already installed.  Installation of this version cannot continue.  To configure or remove the existing version of this product, use Add/Remove Programs on the Control Panel."